refactored code, changed contacts sorting functionality and added updating avatar route logic
