+++ controllers/users/index.js
@@ -1,0 +1,15 @@
+const registration = require("./registration");
+const login = require("./login");
+const logout = require("./logout");
+const current = require("./current");
+const updateSubscription = require("./updateSubscription");
+const updateAvatar = require("./updateAvatar");
+
+module.exports = {
+  registration,
+  login,
+  logout,
+  current,
+  updateSubscription,
+  updateAvatar,
+};
