+++ routes/api/users.js
@@ -1,13 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const ctrlWrapper = require("../../helpers/ctrlWrapper");
-const {
-  registration,
-  login,
-  logout,
-  current,
-  updateSubscription,
-} = require("../../controllers/users");
+const ctrl = require("../../controllers/users");
 const { validate } = require("../../middlewares/validateBody");
 const {
   loginSchema,
@@ -15,16 +9,27 @@ const {
   subscriptionSchema,
 } = require("../../models/user");
 const auth = require("../../middlewares/authMiddleware");
+const upload = require("../../middlewares/uploadMiddleware");
 
-router.post("/register", validate(registerSchema), ctrlWrapper(registration));
-router.post("/login", validate(loginSchema), ctrlWrapper(login));
-router.get("/logout", auth, ctrlWrapper(logout));
-router.get("/current", auth, ctrlWrapper(current));
+router.post(
+  "/register",
+  validate(registerSchema),
+  ctrlWrapper(ctrl.registration)
+);
+router.post("/login", validate(loginSchema), ctrlWrapper(ctrl.login));
+router.get("/logout", auth, ctrlWrapper(ctrl.logout));
+router.get("/current", auth, ctrlWrapper(ctrl.current));
 router.patch(
   "/",
   auth,
   validate(subscriptionSchema),
-  ctrlWrapper(updateSubscription)
+  ctrlWrapper(ctrl.updateSubscription)
+);
+router.patch(
+  "/avatars",
+  auth,
+  upload.single("avatar"),
+  ctrlWrapper(ctrl.updateAvatar)
 );
 
 module.exports = router;
